requirements.txt
torch>=2.2
pytorch-lightning>=2.2
pytorch-forecasting>=1.0
pandas>=2.0
polars>=1.0
numpy
matplotlib
seaborn
mlflow>=3.10
lightgbm>=4.0
snowflake-connector-python
pyarrow
scikit-learn
optuna
jupyter
ipykernel
tabulate

# Phase 1 — TDD + jupytext
pytest>=8.0
pytest-mock>=3.12
pandera>=0.20
jupytext>=1.16
pyyaml>=6.0

# Stage 0 — MLflow 트래킹 + 시각화 강화 (spec: mlflow-experiment-tracking v0.2.0)
plotly>=5.18
